+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 1.0.0
   - Router - стал абстрактным
   - метод `Router::getRoutes()` абстрактный
+  - удалено свойство $routes класса Router
 
 ## 0.0.1
   - initial version